Api_library: rubocop/flog resolved
